requirements.txt
flask>=2.0.0
flask-cors>=3.0.10
flask-limiter>=2.1.0
psycopg2-binary>=2.9.0
redis>=4.0.0
bcrypt>=3.2.0
pyjwt>=2.4.0